requirements.txt
chris_plugin==0.4.0
opencv-python
pydicom
pflog==1.2.26
pftel-client~=1.0.6
# for bug fix on transfer syntax errors
pylibjpeg
pylibjpeg-libjpeg
python-gdcm

# for DICOM image processing
pillow
pytesseract
spacy
nltk